Remove excessive mutex and timer from GameServer

+++ src/services/system_monitor.hpp
@@ -25,7 +25,7 @@ private:
     NamedLogger                     _logger;
     std::unique_ptr<Poco::Timer>    _timer;
 
-    uint32_t                        _lastRSS;
+    size_t                          _lastRSS;
 };
 
 #endif /* system_monitor_hpp */
